requirements.txt
Flask==0.10.1
Flask-Login==0.3.2
Flask-Migrate==1.8.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-WTF==0.12
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
MySQL-python==1.2.5
SQLAlchemy==1.0.12
WTForms==2.1
Werkzeug==0.11.5
alembic==0.8.5
argparse==1.2.1
fake-factory==0.5.7
ipaddress==1.0.16
itsdangerous==0.24
mysqlclient==1.3.7
python-dateutil==2.5.2
python-editor==1.0
six==1.10.0
wsgiref==0.1.2
